2.2.React.createElement(){Elements & the Virtual DOM; Creating React Elements}
Elements & the Virtual DOM

Here's the idea behind React: it's actually really beautiful: we create React element objects that represent HTML elements, like an h1 tag. These don't actually live on the page, they're just objects that describe HTML elements.

But then, we can tell React to look at our element objects, and use them to create real HTML elements on the page, or, on the DOM. This means that we will have a tree of element objects in React and a tree of element objects on the page. To say it a different way, we will have a virtual DOM in React and the real DOM on the page.

And... that's it! Of course, the magic is that, when we change some data on a React element object, React will update the corresponding DOM element automatically.

So... let's create some React elements! To really master this, yea... we're going to do things the hard way first. But, it will be totally worth it.

Creating React Elements

Create a new const el set to React.createElement(). Make this an h2 tag: we're building the title on top of the app. Pass null for the second argument: but this is where you could pass an array of any HTML attributes for the element. These are called 'props' in React - but more on that later. For the third argument, pass whatever you want to put inside, like text: 'Lift History! '.

 Cool! Let's console.log(el): I want you to see that this is just a simple object. Go refresh the page.The element is not, yet, being rendered to the screen in any way. It's just a React element that describes a potential HTML element.