3.3.JSX{JSX & The Babel React Preset; yarn add babel-preset-react --dev}
We need to install a new package: copy the command name to squash the failing in previous commit. Then, paste:

yarn add babel-preset-react --dev
By the way, the next version of this package will be called @babel/preset-react. So, if you see that package name in the future, don't worry, it's really the same thing.

And... done! Try Encore again:

yarn run encore dev --watch
Bah! It fails again - but this is my fault: I forgot to remove my extra ). After removing that, yea! Encore builds successfully! This means that it actually understands our crazy JSX code! Try it - move to your browser and refresh