4.3.React Components{One Component Per File}
One Component Per File

Just like in PHP, we're going to follow a pattern where each React component class lives in its own file.

We only copied the RepLogApp class. But when we pasted, PhpStorm auto-magically added the import for us! Thanks PhpStorm! Gold star!

But, check out this error:
ESLint: React must be in scope when using JSX.

Oh, that's what we just talked about! This is one of those warnings that comes from ESLint.