6.1.Collection & Rendering a Table{class -> className}
The RepLogApp component eventually needs to render all the elements we see on the original app: with a table and a form. No problem! Go find the template for this page: templates/lift/index.html.twig. Hey! There's our table! And the form lives in this included template. Copy the entire old markup. Then, go back to RepLogApp and replace our code with this. But, I don't want to worry about the form yet... so remove that. Oh, and I kinda liked the withHeart feature, so let's make sure we're still printing {heart}.

class -> className

Sweet! Once again, PhpStorm just did an amazing thing - automagically - when we pasted in the code. Check out the className="table table-striped". Hmm, look at the original code in the template: it had class=""... because that's what it should be in HTML! Well... in React, you actually can't use class. You can see it's highlighted:

  Unknown property class found, use className instead

React can't use class because class is a keyword inside of JavaScript. And for that reason, in JSX, you need to use className instead. But ultimately, this will render as a normal class attribute on the page.

And, don't worry, there aren't tons of weird attributes like this in React: this is basically the only one you're likely to use.

The point is: PhpStorm is smart enough to convert our pasted class props to className automatically. Notice that I said props: while we think of these as HTML attributes, they're technically props, which React ultimately renders as attributes.