6.2.Collection & Rendering a Table{Finish our Static "Mock Up"}
Finish our Static "Mock Up"

Let's clean a few things up! We don't need this js-rep-log-table class: it was used by the old JavaScript. And below, this is the column that prints the total weight. Remove the class that was used by the old JavaScript and, for now, just put a TODO inside.

And finally, just to see how it looks with data, let's hack in one fake row full of invented stuff. Use ... for the last column: someday, we'll add a delete link here.

Cool! Building a static version of your app first is a great way to start. And JSX makes that really easy.

Let's go check it out: find your browser and refresh! Hey hey! This is starting to look real!

In our old app, on page load, we make an AJAX request to load the "rep logs"- that's what we call our data - and use that to render the table.

Eventually, we'll do the same thing in React. But before you work with dynamic data, you should first make your app render using static data.