6.3.Collection & Rendering a Table{Rendering a Collection}
Use Static Data First

Check this out, inside render(), create a new constant called repLogs and then set that to some fake data that matches the format of your API. We now have 3 fake rep logs with id, itemLabel and totalWeight.

Rendering a Collection

Below, inside the tbody, we basically want to convert each "rep log" into a tr React element with the data printed inside of it. To do that, we're going to use a really common pattern in React... which might feel a bit weird at first.

Above render(), create a new constant called repLogElement set to repLogs.map(). Pass this a callback function with one argument: repLog. I'll use the arrow syntax for the callback. Inside, we're going to return a React element via JSX: add parenthesis so we can use multiple lines. Then, just build out the row: <tr>, then <td> with {repLog.itemLabel}.

If you're not familiar with the map function, that's ok: it's much less common in PHP. Basically, it loops over each element in repLogs, calls our function, and then, whatever our function returns, is added to the repLogElement array. So, ultimately, repLogElement will be an array of <tr> React element objects.

Add the next <td>. Let's see... ah, this column is "How Many". Fill in the second column with {repLog.reps}, then another <td> with {repLog.totalWeightLifted} and finally one more with ...: this will be the delete link... someday.

Great! Wait... but the tr has a little warning: something about a missing key prop. We'll talk about that in a minute. Until then, ignore that silly warning! What could go wrong?!

Now that we have an array of React element objects, this is pretty sweet: go down, delete the hardcoded row and - wait for it - just print repLogElements.

Yea, it looks a bit crazy: we're literally printing an array of React elements! But, try it - go back to your browser and refresh! It works! It prints each row!

But, we have a big warning from React. It says:

  Warning: each child in an array should have a unique 'key' prop.

Let's fix that next, and introduce a new best practice to keep our code readable.