7.1.The key Prop & Inline Rendering
We just rendered an array of React <tr> element objects. And as a thank you, React has awarded us with a big ugly error! It says:

  Warning: each child in an array should have a unique key prop.

Rude! Here we are, with a perfectly functional table, and React is ruining our magical moment. Well... ok, it's warning us for a good reason. And, PhpStorm was also trying to help.