7.2.The key Prop & Inline Rendering{Why a Collection needs a key}
Why a Collection needs a key

Here's the deal: soon, we're going to use React to perform automagical updates to the UI when data changes. Eventually, we'll use this to add and remove rows to this table as we create and delete rep logs! But... when you have a collection of items like this, if we update the data on one of the rep logs, React isn't totally sure which row to update in the DOM. For example, if this rep log's weight suddenly changed to 150, it's possible that React would update the wrong row!

To help React, we need to give each row a unique key - kind of like how each row in a database table has a primary key. To do this, go to the outer-element of each item and literally add a key prop. This needs to be something that is unique and won't change. So, basically, it should be the id.

Solved! This key prop isn't a big deal, it's just a chore you need to handle each time you render a collection. But don't worry: if you forget, React will remind you!

Try it now: head over and refresh! The page still works, and the warning is gone