7.3.The key Prop & Inline Rendering{Rendering Inline}
Rendering Inline

There's one minor downside to this new setup. Up here, we use the map function to create an array of repLogElements. Down below, we render that.

What's the problem? Well, just that, if you're looking at render() to see your markup, when you see {repLogElements}, you need to scroll back up to see what this is. Whenever possible, it's better to keep all of your markup in one place.

And, we can do that... by being a bit clever. Copy the repLogs.map() code, then delete the repLogElements variable entirely. Back inside JSX, clear out the variable and... paste!

That's it! It's really the same thing we had before! This loops over the repLogs array, builds an array of "rep log" element objects, then... prints them!

Try it! Move over and, refresh! Yes!

At first, the syntax may look weird. But, it's now very obvious that inside the tbody, we are printing a bunch of tr elements.

And if this isn't fancy enough for you, well, you won't be disappointed. If you're still getting used to how the arrow functions look, you may not love this next change. But, if it's too weird, don't use it!