7.4.The key Prop & Inline Rendering{Shorter Arrow Functions}
Shorter Arrow Functions

Because the arrow function is, um, a function, we usually surround the body of the function with curly braces. But, if the only line in your function is the return statement, you can remove the curly braces and just put the code for the return. Oops, I have one extra curly brace.

We now have a function with one argument that returns this JSX element.

Move over and try it! Woohoo! It works! These are the types of little things in React that you don't need to do. And, if it makes your head spin, keep it simple. But, I want you to at least know about these tricks. Because, often, it's these types of shortcuts that end up making React look hard.

Let's finish building the static version of our app next: by adding the form!