8.3.Build the Static App First{Using form defaultValue}
Using form defaultValue

Let's fix the warning:
  Warning: Use the 'defaultValue' or 'value' props on <select> instead of setting 'selected' on <option>.
first before proper styling the form. We'll talk a lot more about forms in React later. But basically, React is a control freak, it really really* wants to manage the values of any elements on your form, including which option is selected. So, instead of using selected="selected", you can use defaultValue="" on the select element and set it to the value of the option you want. I'll skip that part because the first option will be automatically selected anyways.