9.4.State: For Magic Updating Good Times{But... who Updates State?}
But... who Updates State?

Of course, our ultimate goal is not just to update state via the cool React dev tools. Nope, we want to update the state when the user clicks a row.

Before we do that, there is one small housekeeping item. In addition to destructuring state, I like to do the same with props. Add const { withHeart } = this.props. Then below, use if withHeart.

It's a small detail, but it's nice to setup all the variables right on top.

Now, let's add some click magic and update our state!