10.2.Handling Events (like onClick) ! {Updating with this.setState()}
Updating with this.setState()

Cool. Let's review our goal: to highlight a row when we click on it. So... hmm... onClick: if we could update the highlightedRowId state to the correct id, React would re-render and take care of the rest! Easy! Inside the arrow function, update the state with this.setState(). Pass this an object with the state key or keys that you want to change. For us, highlightedRowId set to the id of this rep log: repLog.id.

Coolio! But, an important note! In the constructor, we initialized the state by setting the this.state property directly. This is the only place, ever, that you will change or set the state property directly. Everywhere else, always, you need to call this.setState(). If you don't, puppies will stare at you with sad eyes.

And, more important, if you modify the state property directly, React won't re-render. The reason is simple: this is what React uses to know that you changed the state and so, to start the re-rendering.

Bah, let's go try it already! Refresh! And... click! Woohoo! We just added our first bit of interactivity. In the React dev tools, if you click on RepLogApp, you can watch the highlightedRowId state change as we click the rows. Pretty freaking cool.